asic: bm1366: Use the lowest possible pll frequency
Keep on looping to find alternative divider settings which still result
in the requested target hash frequency. The one that results in the
lowest pll frequency (25 * fbdiv / refdiv) is used. This is more in line
with the values used in the raw commands in do_frequency_ramp_up().

Since the resulting values ware a bit lower than the ones used by the
raw commands in do_frequency_ramp_up(), the allowed minimal value for fbdiv
had to be increased from 144 to 160. After doing so, using this function
gives the exact same results as the raw commands in do_frequency_ramp_up().

Signed-off-by: Robin van der Gracht <robinvandergracht@gmail.com>